requirements.txt
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.3
matplotlib==2.1.2
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.0
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
scikit-learn==0.19.1
scipy==1.0.0
bleach==1.5.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==1.5.1
psutil==5.4.5
h5py==2.7.1
